requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
openai>=1.30.0
chromadb>=1.5.5
exa-py>=2.10.0
tiktoken>=0.7.0
sse-starlette>=2.1.0
pydantic>=2.0.0
httpx>=0.27.0
